separated clothing and img box
