+++ src/App.js
@@ -7,7 +7,9 @@ import data from './data';
 import 'bootstrap/dist/css/bootstrap.css';
 import { Button, Nav, Navbar, NavDropdown, MenuItem, NavItem } from 'react-bootstrap';
 import { LinkContainer } from 'react-router-bootstrap';
-import {Services, Front} from './Services.js';
+import {Services, Front, Back, Management} from './Services.js';
+import {Product, Clothing} from './Clothing.js';
+import {ImageBox} from './ImageBox.js'
 
 
 var App = React.createClass({
@@ -43,8 +45,8 @@ var Header = React.createClass({
                   <span className="caret"></span>
                 </Link> 
               </li>
-              <li key="codersCloths">
-                <Link to="codersCloths">Coders Cloths
+              <li key="Clothing">
+                <Link to="Clothing">Coders Cloths
                 </Link>
               </li>
               <li key="about">
@@ -63,46 +65,6 @@ var Header = React.createClass({
 })
         
 
-
-var CodersCloths = React.createClass({
-	getInitialState(){
-		return (
-			{product:null}
-			)
-	},
-	componentDidMount(){
-		this.setState({product: data.getproduct()})
-	},
-	render: function() {
-  	var arr= [];
- 
-  	for(var product in this.state.product){
-      console.log(this.state.product.product1[0].price)
-  		arr.push(<h2>{product}</h2>)
-  		this.state.product[product].map((item,i)=> {
-  		arr.push(<Product key={item.id} name={item.name} price={item.price} />)
-     		})
-     	}
-    return (
-      <div className="flex-container col-md-4">
-  		  {arr}
-      </div>
-    )
-  }
-})
-var Product = React.createClass({
-	getInitialState(){
-		return ({product: null})
-	},
-	render: function() {
-    return (
-      <ul className="box">
-    		<li key={this.props.keyName}>{this.props.name}</li>
-    		<li> {this.props.price}</li>
-      </ul>
-    )
-  }
-})
 
 var Contact = React.createClass({
   getInitialState(){
@@ -172,7 +134,7 @@ ReactDOM.render(
           <Route path="Services" component={Services}>  
             <Route path="Services/front" component={Front} />
           </Route>    
-          <Route path="codersCloths" component={CodersCloths} />
+          <Route path="Clothing" component={Clothing} />
           <Route path="about" component={About} />
           <Route path="contact" component={Contact} />
     </Route>
