+++ src/ImageBox.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+var ImageBox = React.createClass({
+	render: function() {
+		return (
+			<div>
+				<h1>side scrolling images goes here</h1>
+			</div>
+		)
+	}
+})
+
+export {ImageBox};
